requirements.txt
streamlit>=1.30
torch>=2.1.0
transformers>=4.37.0
sympy>=1.13.1
pillow>=10.0.0
matplotlib>=3.8.0
peft~=0.5.0
safetensors
